// UI/PA01/src/app/models/estate.ts
export class estate{
    EstateId: string;
    EstateName: string;
    EstatePrice: string;
    DiscountPrice: string;
    Clicks: number;
    IsActive: string;
    IsDiscount: string;
    EstateAdress: string;
    EstateDescription: string;
    EstateLocationId: string;
    EstateRegDate: string;
    EstateEditDate: string;
    EstateUserAmount: number;
    EstateIMG1: string;
    EstateIMG2: string;
    EstateIMG3: string;
    EstateIMG4: string;
}